# siteripper/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScrapeInput {
  url: string;
  maxPages?: number;
}

export type ScrapeJobStatus = typeof ScrapeJobStatus[keyof typeof ScrapeJobStatus];


export const ScrapeJobStatus = {
  pending: 'pending',
  running: 'running',
  done: 'done',
  error: 'error',
} as const;

export interface ScrapeJob {
  jobId: string;
  status: ScrapeJobStatus;
  url: string;
  pagesFound: number;
  assetsFound: number;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}