# siteripper/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scrape
    description: Site scraping operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /scrape/start:
    post:
      operationId: startScrape
      tags: [scrape]
      summary: Start a scrape job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeInput"
      responses:
        "200":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeJob"
        "400":
          description: Bad request
  /scrape/status/{jobId}:
    get:
      operationId: getScrapeStatus
      tags: [scrape]
      summary: Get job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeJob"
        "404":
          description: Job not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ScrapeInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        maxPages:
          type: integer
    ScrapeJob:
      type: object
      required: [jobId, status, url, pagesFound, assetsFound]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, running, done, error]
        url:
          type: string
        pagesFound:
          type: integer
        assetsFound:
          type: integer
        message:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]