# siteripper/lib/api-zod/src/generated/types/scrapeJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapeJobStatus } from './scrapeJobStatus';

export interface ScrapeJob {
  jobId: string;
  status: ScrapeJobStatus;
  url: string;
  pagesFound: number;
  assetsFound: number;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}